requirements.txt
fastapi~=0.116.1
uvicorn
httpx
python-dotenv~=1.1.1
fastapi-cache2
slowapi

pydantic~=2.11.7